+++ spec/dionysus/topic_name_spec.rb
@@ -4,7 +4,7 @@ require "spec_helper"
 
 RSpec.describe Dionysus::TopicName do
   describe "#to_s" do
-    subject(:to_s) { described_class.new(namespace, name).to_s }
+    subject(:topic_to_s) { described_class.new(namespace, name).to_s }
 
     let(:namespace) { "v3" }
     let(:name) { "rentals" }