+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 ## [Unreleased]
 
-## [0.1.0] - 2023-10-18
+## [0.1.0] - 2023-12-18
 
 - Turn private gem `bookingsync-prometheus` into `dionysus-rb` and release it publicly
 - Allow to customize `consumer_group_prefix`
+- Support Ruby 3.2
 
 
 Migration path from `bookingsync-prometheus`: